refactor, to produce separate control for adding or editing
